Update node-cardano-wallet dependency
This dep is only used for testing due to Jest restrictions, and it
indirectly uses a very old version of `rust-cardano`. The dep now points
to an experimental branch that has a version of `js-cardano-wasm` that
supports testnet address generation and also has an up to date version
of `rust-cardano`, which is likely what was causing the issues in
`checkAddress()`.